requirements.txt
Django==1.6.5
South==0.8.4

py==1.4.20
pytest==2.5.2
pytest-cov==1.6
pytest-django==2.6
pytest-flakes==0.2
pytest-pep8==1.0.5
factory-boy==2.2.1
mock==1.0.1

fabric==1.8.3

psycopg2==2.5.1